requirements.txt.txt
pillow==10.3.0
pywin32==306